+++ static/popups/correlations/correlationsUtils.jsx
@@ -32,6 +32,7 @@ function createScatter(ctx, data, xProp, yProp, onClick) {
   delete scatterCfg.options.scales.xAxes[0].ticks;
   delete scatterCfg.options.scales.yAxes[0].ticks;
   scatterCfg.options.onClick = onClick;
+  scatterCfg.options.maintainAspectRatio = false;
   // eslint-disable-next-line new-cap
   const chart = chartUtils.createChart(ctx, scatterCfg);
   return chart;