+++ static/popups/window/instances_popup_main.jsx
@@ -5,6 +5,8 @@ import "../../adapter-for-react-16";
 import app from "../../reducers/dtale";
 import Instances from "../Instances";
 
+require("../../dtale/DataViewer.css");
+
 const dataId = app.getHiddenValue("data_id");
 
 ReactDOM.render(<Instances dataId={dataId} iframe={true} />, document.getElementById("popup-content"));